+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.2.0] - 2025-09-06
+
 ### Added
 
 - New course data
@@ -14,6 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changes
 
 - Update dependencies
+- Fix selected course save by changning trigger from beforunload to visibilitychange
+- Handle corrupted localstorage
 
 ## [1.1.0] - 2022-04-04
 
@@ -93,7 +97,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Timetable of selected courses.
 - Display individual and total credit of selected courses.
 
-[unreleased]: https://github.com/kilicbaran/boun-course-planner/compare/v1.0.0...HEAD
+[unreleased]: https://github.com/kilicbaran/boun-course-planner/compare/v1.2.0...HEAD
+[1.2.0]: https://github.com/kilicbaran/boun-course-planner/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/kilicbaran/boun-course-planner/compare/v1.0.2...v1.1.0
 [1.0.2]: https://github.com/kilicbaran/boun-course-planner/compare/v1.0.1...v1.0.2
 [1.0.1]: https://github.com/kilicbaran/boun-course-planner/compare/v1.0.0...v1.0.1